requirements.txt
ffmpeg-python==0.2.0
future==1.0.0
numpy>=1.21.0
opencv-python>=4.5.0
scikit-image>=0.19.0
pillow>=10.0.0
google-genai>=0.1.0
supabase>=2.0.0
python-dotenv>=1.0.0
